refactor train_model for add more model modules
