+++ src/models/models/__init__.py
@@ -1,0 +1,3 @@
+__all__ = [
+    "isolationForest",
+]
