+++ src/models/models/isolationForest/__init__.py
@@ -1,0 +1,5 @@
+__all__ = [
+    "preprocessing",
+    "train",
+    "plotting"
+]
